Add new service and controller tests

+++ src/main/java/com/sparsis/libraryapi/service/BookServiceImpl.java
@@ -4,6 +4,8 @@ import com.sparsis.libraryapi.exception.BusinessException;
 import com.sparsis.libraryapi.model.entity.Book;
 import com.sparsis.libraryapi.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -32,12 +34,23 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id) throws IllegalArgumentException {
+        if(id == null) {
+            throw new IllegalArgumentException("ID cannot be null");
+        }
         repository.deleteById(id);
     }
 
     @Override
     public Book update(Book book) {
+        if(book == null || book.getId() == null) {
+            throw new IllegalArgumentException("ID cannot be null");
+        }
         return repository.save(book);
+    }
+
+    @Override
+    public Page<Book> find(Book book, Pageable pageRequest) {
+        return null;
     }
 }
